Agri app/requirements.txt
streamlit>=1.22
pillow>=9.5
google-genai>=0.0.4
python-dotenv>=1.0.0
